requirements.txt
# Application
Flask==0.12.2
Flask-SQLAlchemy==2.2
jsonschema==2.6.0
gunicorn==19.7.1

# Testing
nose==1.3.7
